// src/scripts/external/typescript/jquery.d.ts
interface BaseJQueryEventObject extends Event {
  /**
   * The current DOM element within the event bubbling phase.
   * @see {@link https://api.jquery.com/event.currentTarget/}
   */
  currentTarget: Element;
  /**
   * An optional object of data passed to an event method when the current executing handler is bound.
   * @see {@link https://api.jquery.com/event.data/}
   */
  data: any;
  /**
   * The element where the currently-called jQuery event handler was attached.
   * @see {@link https://api.jquery.com/event.delegateTarget/}
   */
  delegateTarget: Element;
  /**
   * Returns whether event.preventDefault() was ever called on this event object.
   * @see {@link https://api.jquery.com/event.isDefaultPrevented/}
   */
  isDefaultPrevented(): boolean;
  /**
   * Returns whether event.stopImmediatePropagation() was ever called on this event object.
   * @see {@link https://api.jquery.com/event.isImmediatePropagationStopped/}
   */
  isImmediatePropagationStopped(): boolean;
  /**
   * Returns whether event.stopPropagation() was ever called on this event object.
   * @see {@link https://api.jquery.com/event.isPropagationStopped/}
   */
  isPropagationStopped(): boolean;
  /**
   * The namespace specified when the event was triggered.
   * @see {@link https://api.jquery.com/event.namespace/}
   */
  namespace: string;
  /**
   * The browser's original Event object.
   * @see {@link https://api.jquery.com/category/events/event-object/}
   */
  originalEvent: Event;
  /**
   * If this method is called, the default action of the event will not be triggered.
   * @see {@link https://api.jquery.com/event.preventDefault/}
   */
  preventDefault(): any;
  /**
   * The other DOM element involved in the event, if any.
   * @see {@link https://api.jquery.com/event.relatedTarget/}
   */
  relatedTarget: Element;
  /**
   * The last value returned by an event handler that was triggered by this event, unless the value was undefined.
   * @see {@link https://api.jquery.com/event.result/}
   */
  result: any;
  /**
   * Keeps the rest of the handlers from being executed and prevents the event from bubbling up the DOM tree.
   * @see {@link https://api.jquery.com/event.stopImmediatePropagation/}
   */
  stopImmediatePropagation(): void;
  /**
   * Prevents the event from bubbling up the DOM tree, preventing any parent handlers from being notified of the event.
   * @see {@link https://api.jquery.com/event.stopPropagation/}
   */
  stopPropagation(): void;
  /**
   * The DOM element that initiated the event.
   * @see {@link https://api.jquery.com/event.target/}
   */
  target: Element;
  /**
   * The mouse position relative to the left edge of the document.
   * @see {@link https://api.jquery.com/event.pageX/}
   */
  pageX: number;
  /**
   * The mouse position relative to the top edge of the document.
   * @see {@link https://api.jquery.com/event.pageY/}
   */
  pageY: number;
  /**
   * For key or mouse events, this property indicates the specific key or button that was pressed.
   * @see {@link https://api.jquery.com/event.which/}
   */
  which: number;
  /**
   * Indicates whether the META key was pressed when the event fired.
   * @see {@link https://api.jquery.com/event.metaKey/}
   */
  metaKey: boolean;
}

interface JQueryInputEventObject extends BaseJQueryEventObject {
  altKey: boolean;
  ctrlKey: boolean;
  metaKey: boolean;
  shiftKey: boolean;
}

interface JQueryMouseEventObject extends JQueryInputEventObject {
  button: number;
  clientX: number;
  clientY: number;
  offsetX: number;
  offsetY: number;
  pageX: number;
  pageY: number;
  screenX: number;
  screenY: number;
}

interface JQueryKeyEventObject extends JQueryInputEventObject {
  char: any;
  charCode: number;
  key: any;
  keyCode: number;
}

export interface JQueryEventObject
  extends BaseJQueryEventObject,
  JQueryInputEventObject,
  JQueryMouseEventObject,
  JQueryKeyEventObject { }